[ethers-v5] refactor: code split the event methods
This commit creates an events.ts file and moves the helper functions
which were at the bottom at the codegen/index file to the new file
